# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
tags:
  - name: Users
  - name: Auth
  - name: league
  - name: Teams
  - name: Games
paths:
  /Register:
    post:
      tags:
        - Auth
      summary: Add new user
      description: A new user from the client is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: A new user has been added
        "400":
          description: wrong input parameters
  /Login:
    post:
      tags:
        - Auth
      summary: Login an exist user
      description: An exist user login to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginScheme"
      responses:
        "200":
          description: >
            User was found in the DB and authenitcated successfully.
            The session ID is returned in a cookie named `session`. 
            You should include this cookie in subsequent requests
          # headers:
          #   Set-Cookie:
          #     schema:
          #       type: string
          #       example: session=87s97sdhgf279293f8w0e9
            # Note that the Set-Cookie header and securitySchemes are not connected in any way, and the Set-Header definition is for documentation purposes only.
        "401":
          description: Authentication failed
  /logout:
    post:
      tags:
        - Auth
      summary: logout a user from the system and delete his cookie
      responses:
        "200":
          description: "the user has ben successfully logged out"
  /league/getDetails:
    get:
      tags:
        - league
      summary: Return the details about the current league
      responses:
        "200":
          description: data about the current league returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
        "400":
          description: wrong input parameters

  /games/addGame:
    post:
      tags:
        - Games
      summary: A union manager from client is add game to league
      description: a JSON object contains gameId to add to league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gamedata'
        description: add new game
      responses:
        '200':
          description: A new game has been added
        '400':
          description: Wrong input parameters

  /games/addEvent:
    post:
      tags:
        - Games
      summary: A union manager from client is add game to league
      description: a JSON object contains gameId to add to league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/eventslog'
        description: add new event
      responses:
        '200':
          description: A new event has been added
        '400':
          description: Wrong input parameters
          
  /users/favoritePlayers:
    get:
      tags:
        - Users
      summary: Return the favorite players that were saved by user
      parameters:
      responses:
        "200":
          description: data about the user's favorites players returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPreview"
        "401":
          description: user is not authorized
    post:
      summary: save player in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId:
                  type: string
                  description: player's id which the user saved
                  example: 85
      responses:
        "201":
          description: The player successfully saved as favorite
        "401":
          description: user is not authorized

  /users/favoriteGames:
    get:
      tags:
        - Users
      summary: Return the favorite games that were saved by user
      parameters:
      responses:
        "200":
          description: data about the user's favorites games returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPrevie"
        "401":
          description: user is not authorized
    post:
      summary: save game in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                gameId:
                  type: string
                  description: game's id which the user saved
                  example: 939
      responses:
        "201":
          description: The game successfully saved as favorite
        "401":
          description: user is not authorized

  /users/favoriteTeams:
    get:
      tags:
        - Users
      summary: Return the favorite teams that were saved by user
      parameters:
      responses:
        "200":
          description: data about the user's favorites teams returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "401":
          description: user is not authorized
    post:
      summary: save team in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                teamsId:
                  type: string
                  description: teams's id which the user saved
                  example: 123
      responses:
        "201":
          description: The game successfully saved as favorite
        "401":
          description: user is not authorized

  /teams/teamFullDetails/{teamId}:
    get:
      summary: get full deatils about a team
      description: useful for team page
      tags:
        - Teams
      parameters:
        - in: path
          name: teamId
          description: team's id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: data about the team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: teamId is not exists
        
components:
  schemas:
    User:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins between 3 to 8 letters only
          example: ladygaga
        firstname:
          type: string
          description: user's first name
          example: Stefani
        lastname:
          type: string
          description: user's last name
          example: Germanotta
        country:
          type: string
          description: user's country
          example: USA
        password:
          maxLength: 10
          minLength: 5
          type: string
          description:
            password conatins between 5 to 10 chars, at least one of them
            is a number and a special char
          format: password
          example: lady@56
        email:
          pattern: email
          type: string
          description: user's email
          example: ladygaga@gmail.com
        imageUrl:
          pattern: URL
          type: string
          description: URL for the user's image
          example: https://cloudinary.com/
      description: Model containing user info
    LoginScheme:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          description: username conatins letters only
          example: ladygaga
        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          description: password conatins at least one number and one special char
          example: lady@56
    League:
      type: object
      properties:
        league_name:
          type: string
          description: league name
          example: Superliga
        current_season_name:
          type: string
          description: current season name
          example: "2020/2021"
        current_stage_name:
          type: string
          description: current stage name
          example: Championship Round
        NextGame:
          type: object
          description: Contains next game details
          properties:
            date:
              type: string
              example: 3/4/20201
              description: date
            home_team:
              type: string
              example: liverpool
              description: the home team
            away_team:
              type: string
              example: everton
              description: the away team
    PlayerPreview:
      type: object
      properties:
        name:
          type: string
          description: player's name
          example: Sten Michael Grytebust
        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png
        position:
          type: integer
          description: player's position in team
          example: 1
        team_name:
          type: string
          description: player's team name
          example: København
    Team:
      type: object
      properties:
        name:
          type: string
          description: team name
          example: København
        # coach:
        #   type: string
        #   description: coach name
        #   example: tommy
        players:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/PlayerPreview"
    GameDetails:
      type: array
      items:
        type: object
        description: Contains past games
        properties:
          date:
            type: string
            example: 3/3/2020
            description: date
          time:
            type: string
            example: 20:30
            description: date
          hometeam:
            type: string
            example: liverpool
            description: the home team
          awayteam:
            type: string
            example: manchester
            description: the away team
          field:
            type: string
            example: Anfield
            description: the stadium


    # futureGames:
    #   type: array
    #   items:
    #     type: object
    #     description: Contains future games
    #     properties:
    #       score:
    #         type: string
    #         example: 2-1
    #         description: score game
    #       date:
    #         type: string
    #         example: 3/3/2020
    #         description: date
    #       opponent:
    #         type: string
    #         example: liverpool
    #         description: the opponent team

    coachdata:
      type: object
      description: Contains data on coach
      properties:
        name:
          type: string
          example: Pep
          description: coach name
        team:
          type: string
          example: juventus
          description: coach teams name
        picture:
          type: string
          format: URL
          example: https://cloudinary.com/juventus
          description: coach image
    
    gamedata:
      type: object
      description: Contains data on new game
      properties:
        gameID:
          type: integer
          example: 231
          description: the id of the game
        gamedate:
          type: string
          example: 26-12-2015
          descriptio/n: the date of the game
        gametime:
          type: string
          example: '15:45:20'
          description: the time of the game
        hometeamID:
          type: integer
          example: 123
          description: id of home team
        awayteamID:
          type: integer
          example: 456
          description: id of away team
        field:
          type: string
          example: Bloomfield
          description: field
        score:
          type: string
          example: 1-1
          description: score

    eventslog:
      type: object
      description: Contains game events
      properties:
        gameID:
          type: integer
          example: 231
          description: the id of the game
        eventdate:
          type: string
          example: 3-4-2020
          description: date
        eventtime:
          type: string
          example: 19:30:20
          description: time
        eventminute:
          type: integer
          example: 10
          description: the minute of the game
        dataevent:
          type: string
          example: 10
          description: goal
        playerID:
          type: integer
          example: 145
          description: the id of the player who did the event